SOF-129: Code structure refactor for consistency and clarity
