+++ app/src/main/java/com/vci/vectorcamapp/location/di/LocationModule.kt
@@ -1,0 +1,20 @@
+package com.vci.vectorcamapp.location.di
+
+import com.vci.vectorcamapp.location.data.repository.LocationRepositoryImplementation
+import com.vci.vectorcamapp.location.domain.repository.LocationRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class LocationModule {
+
+    @Binds
+    @Singleton
+    abstract fun bindLocationRepository(
+        impl: LocationRepositoryImplementation
+    ): LocationRepository
+}
