+++ app/src/main/java/com/vci/vectorcamapp/location/domain/repository/LocationRepository.kt
@@ -1,0 +1,7 @@
+package com.vci.vectorcamapp.location.domain.repository
+
+import android.location.Location
+
+interface LocationRepository {
+    suspend fun getCurrentLocation(): Location
+}
